# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "RAW.Ninja is an application to store RAW image files.",
        "title": "RAW.Ninja API",
        "contact": {},
        "version": "0.1"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/photos": {
            "get": {
                "description": "Returns all photo descriptors for the current user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "photos"
                ],
                "summary": "List user's photo descriptors endpoint",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/photo.Response"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/common.StatusMessage"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/common.StatusMessage"
                        }
                    }
                }
            }
        },
        "/photos/:id": {
            "get": {
                "description": "Returns the photo descriptor with the provided ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "photos"
                ],
                "summary": "Get photo endpoint",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID of the photo information to collect",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/photo.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/common.StatusMessage"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/common.StatusMessage"
                        }
                    }
                }
            },
            "put": {
                "description": "Updates tags and favorite setting for RAW file",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "photos"
                ],
                "summary": "Update photo endpoint for tags and favorite setting",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID of the photo to update",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "The new version of the photo to use for update",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/photo.Response"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/common.StatusMessage"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/common.StatusMessage"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/common.StatusMessage"
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes the photo with the provided ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "photos"
                ],
                "summary": "Delete photo endpoint",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID of the photo to delete",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/photo.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/common.StatusMessage"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/common.StatusMessage"
                        }
                    }
                }
            }
        },
        "/photos/:id/raw": {
            "get": {
                "description": "Returns the RAW file for the provided ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "photos"
                ],
                "summary": "Download RAW file endpoint",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID of the RAW photo to download",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "integer"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/common.StatusMessage"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/common.StatusMessage"
                        }
                    }
                }
            }
        },
        "/photos/:id/thumbnail": {
            "get": {
                "description": "Returns the thumbnail for the provided ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "photos"
                ],
                "summary": "Thumbnail image endpoint",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID of the thumbnail to download",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "integer"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/common.StatusMessage"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/common.StatusMessage"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "common.StatusMessage": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "descriptor.Response": {
            "type": "object",
            "properties": {
                "favorite": {
                    "type": "boolean"
                },
                "filename": {
                    "type": "string"
                },
                "format": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "metadata": {
                    "$ref": "#/definitions/image.Response"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "thumbnail": {
                    "type": "string"
                },
                "thumbnail_height": {
                    "type": "integer"
                },
                "thumbnail_width": {
                    "type": "integer"
                },
                "uploaded": {
                    "type": "string"
                }
            }
        },
        "image.PresignedRequest": {
            "type": "object",
            "properties": {
                "method": {
                    "type": "string"
                },
                "mode": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "image.Response": {
            "type": "object",
            "properties": {
                "ISO": {
                    "type": "integer"
                },
                "aperture": {
                    "type": "number"
                },
                "camera_make": {
                    "type": "string"
                },
                "camera_model": {
                    "type": "string"
                },
                "camera_sw": {
                    "type": "string"
                },
                "colors": {
                    "type": "integer"
                },
                "data_size": {
                    "type": "integer"
                },
                "height": {
                    "type": "integer"
                },
                "lens_make": {
                    "type": "string"
                },
                "lens_model": {
                    "type": "string"
                },
                "shutter": {
                    "type": "number"
                },
                "timestamp": {
                    "type": "string"
                },
                "width": {
                    "type": "integer"
                }
            }
        },
        "photo.Response": {
            "type": "object",
            "properties": {
                "descriptor": {
                    "$ref": "#/definitions/descriptor.Response"
                },
                "id": {
                    "type": "string"
                },
                "raw": {
                    "$ref": "#/definitions/image.PresignedRequest"
                },
                "thumbnail": {
                    "$ref": "#/definitions/image.PresignedRequest"
                }
            }
        },
...
    },
...
}